move duration code
