+++ modules/overrides/userconfigurable/client/limits.go
@@ -1,8 +1,6 @@
 package client
 
 import (
-	"encoding/json"
-	"fmt"
 	"time"
 
 	filterconfig "github.com/grafana/tempo/pkg/spanfilter/config"
@@ -28,44 +26,6 @@ func (l *Limits) GetMetricsGenerator() *LimitsMetricsGenerator {
 	}
 	return nil
 }
-
-type Duration struct {
-	time.Duration
-}
-
-// MarshalJSON implements json.Marshaler.
-func (d *Duration) MarshalJSON() ([]byte, error) {
-	return []byte(fmt.Sprintf(`"%s"`, d.String())), nil
-}
-
-// UnmarshalJSON implements json.Unmarshaler.
-func (d *Duration) UnmarshalJSON(input []byte) error {
-	var unmarshalledJSON interface{}
-
-	err := json.Unmarshal(input, &unmarshalledJSON)
-	if err != nil {
-		return err
-	}
-
-	switch value := unmarshalledJSON.(type) {
-	case float64:
-		d.Duration = time.Duration(value)
-	case string:
-		d.Duration, err = time.ParseDuration(value)
-		if err != nil {
-			return err
-		}
-	default:
-		return fmt.Errorf("invalid duration: %#v", unmarshalledJSON)
-	}
-
-	return nil
-}
-
-var (
-	_ json.Marshaler   = (*Duration)(nil)
-	_ json.Unmarshaler = (*Duration)(nil)
-)
 
 type LimitsMetricsGenerator struct {
 	Processors         listtomap.ListToMap `json:"processors,omitempty"`
